Removes old QITE files.
